+++ torchvision/csrc/io/image/cpu/read_image_impl.h
