+++ torchvision/csrc/io/image/cpu/read_write_file_impl.cpp
@@ -1,4 +1,4 @@
-#include "read_write_file_cpu.h"
+#include "read_write_file_impl.h"
 
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
