+++ torchvision/csrc/io/image/cpu/read_write_file_impl.h
