+++ torchvision/csrc/io/image/cpu/readpng_impl.h
