+++ torchvision/csrc/io/image/cpu/writejpeg_impl.h
