+++ torchvision/csrc/io/image/cpu/writepng_impl.cpp
@@ -1,4 +1,4 @@
-#include "writejpeg_cpu.h"
+#include "writejpeg_impl.h"
 
 #if !PNG_FOUND
