+++ torchvision/csrc/io/image/cpu/writepng_impl.h
